requirements/python-base.txt
numpy==1.19.4
scikit-learn==0.23.2
tensorflow==2.3.1
keras==2.4.3
dvc[all]==1.10.1
pillow==8.0.1
mlflow==1.12.1
